[Change] js-mouseFollow-Center  Changed the value of the transition property when hovering so that the string "CLICK" can be displayed. Added a hidden area for this purpose.
- transition property
Delayed processing of the center part by transform was done in style.css, so we improved it so that the style is specified in JS. overwriting by style is prevented.

- Display the "CLICK" string.
Expanded the hover area to "topMainvisual" in consideration of the future implementation of video modal that allows video playback.
Along with that, header and newsticker hide the mouse stalker.

- Added a style to hide the mouse stalker.