[Change] We made the class naming capitalize.
Class naming is done from the design stage and is not something that should be changed at this stage. However, as the number of layers increases and the naming conventions become more diverse, the readability of the classes decreases.

Therefore, we changed the class names only for the modules centered on the Site layer.